requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
pydantic
Pydantic-Settings